Pass through $config throught, move functions to LegacyHelpers subnamespace
LegacyHelpers are not referenced from anything else in src/; this means that
they're now explicitly calling the constructors. These calls silently ignored
the provided configuration, and used the default one instead. Directly calling
the constructors means that it's always passed down the stack, and most classes
now require to have a config object passed in.

fixes #3